Remove SettingsPanel component
