+++ gui_dev/src/components/Settings/Settings.jsx
@@ -1,20 +1,44 @@
-import { SettingsPanel } from "./SettingsPanel";
+import { Switch } from "@/components";
 import { useSettingsStore } from "@/stores";
 import styles from "./Settings.module.css";
 
-export const Settings = ({ settingsKey }) => {
-  const { settings } = useSettingsStore((state) => ({
+export const Settings = () => {
+  const { settings, updateSettings } = useSettingsStore((state) => ({
     settings: state.settings,
+    updateSettings: state.updateSettings,
   }));
 
   if (!settings) {
     return <div>Loading settings...</div>;
   }
 
+  const handleChange = (featureKey, isEnabled) => {
+    console.log(settings.features);
+
+    const updatedSettings = {
+      ...settings,
+      features: {
+        ...settings.features,
+        [featureKey]: isEnabled,
+      },
+    };
+
+    updateSettings(updatedSettings);
+  };
+
   return (
-    <div className={styles.settingsContainer}>
-      <h2>Settings</h2>
-      <SettingsPanel settingsKey={settingsKey}/>
+    <div className={styles.settingsPanel}>
+      <h2>Features</h2>
+      <div className="feature-list">
+        {Object.entries(settings.features).map(([key, value]) => (
+          <Switch
+            key={key}
+            label={key}
+            isEnabled={value}
+            onChange={(isEnabled) => handleChange(key, isEnabled)}
+          />
+        ))}
+      </div>
     </div>
   );
 };
